requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
numpy==1.18.5
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
redis==3.5.3
pymodm==0.4.3
PyYAML==5.3.1
mypy==0.782
flake8==3.8.3
black==19.10b9
bandit==1.6.2
